+++ src/filesystem/TankFile.cpp
@@ -11,7 +11,7 @@
 //
 // ================================================================================================
 
-#include "tank_file.hpp"
+#include "TankFile.hpp"
 #include <cmath> // For std::ceil()
 #include <filesystem>
 #include <cassert>
