+++ src/state/test/TankTestState.cpp
@@ -7,7 +7,7 @@
 #include <spdlog/spdlog.h>
 
 #include "cfg/IConfig.hpp"
-#include "filesystem/tank_file.hpp"
+#include "filesystem/TankFile.hpp"
 
 namespace fs = std::filesystem;
